refactored data generatoring for general discounts
