+++ database-mongodb/priceAndPromoSampleData.js
@@ -1,1 +1,1 @@
-[{"product_id":1,"price":"20.00","discount":"6","start":"2020-08-22T02:10:58.255Z","expiry":"2020-09-16T02:10:58.255Z"},{"product_id":2,"price":"86.00","discount":"2","start":"2020-08-20T19:35:14.837Z","expiry":"2020-09-14T19:35:14.837Z"},{"product_id":3,"price":"19.00","discount":"9","start":"2020-08-19T01:34:17.919Z","expiry":"2020-09-13T01:34:17.919Z"},{"product_id":4,"price":"77.00","discount":"8","start":"2020-08-18T22:15:20.894Z","expiry":"2020-09-12T22:15:20.894Z"},{"product_id":5,"price":"39.00","discount":"7","start":"2020-08-19T10:26:05.532Z","expiry":"2020-09-13T10:26:05.532Z"},{"product_id":6,"price":"15.00","discount":"4","start":"2020-08-22T04:34:32.490Z","expiry":"2020-09-16T04:34:32.490Z"},{"product_id":7,"price":"59.00","discount":"5","start":"2020-08-18T04:51:28.162Z","expiry":"2020-09-12T04:51:28.162Z"},{"product_id":8,"price":"91.00","discount":"4","start":"2020-08-21T17:05:22.113Z","expiry":"2020-09-15T17:05:22.113Z"},{"product_id":9,"price":"93.00","discount":"1","start":"2020-08-19T07:46:54.430Z","expiry":"2020-09-13T07:46:54.430Z"},{"product_id":10,"price":"56.00","discount":"2","start":"2020-08-22T02:14:30.127Z","expiry":"2020-09-16T02:14:30.127Z"},{"product_id":11,"price":"65.00","discount":"9","start":"2020-08-17T15:19:51.891Z","expiry":"2020-09-11T15:19:51.891Z"},{"product_id":12,"price":"25.00","discount":"9","start":"2020-08-20T01:48:40.062Z","expiry":"2020-09-14T01:48:40.062Z"},{"product_id":13,"price":"85.00","discount":"2","start":"2020-08-22T02:52:07.492Z","expiry":"2020-09-16T02:52:07.492Z"},{"product_id":14,"price":"67.00","discount":"9","start":"2020-08-20T08:45:15.977Z","expiry":"2020-09-14T08:45:15.977Z"},{"product_id":15,"price":"80.00","discount":"2","start":"2020-08-20T14:44:51.006Z","expiry":"2020-09-14T14:44:51.006Z"},{"product_id":16,"price":"15.00","discount":"1","start":"2020-08-19T04:59:48.108Z","expiry":"2020-09-13T04:59:48.108Z"},{"product_id":17,"price":"81.00","discount":"5","start":"2020-08-22T08:27:07.679Z","expiry":"2020-09-16T08:27:07.679Z"},{"product_id":18,"price":"98.00","discount":"8","start":"2020-08-22T02:17:53.444Z","expiry":"2020-09-16T02:17:53.444Z"},{"product_id":19,"price":"63.00","discount":"8","start":"2020-08-19T12:02:04.112Z","expiry":"2020-09-13T12:02:04.112Z"},{"product_id":20,"price":"11.00","discount":"3","start":"2020-08-20T18:35:44.508Z","expiry":"2020-09-14T18:35:44.508Z"},{"product_id":21,"price":"74.00","discount":"3","start":"2020-08-19T04:31:41.067Z","expiry":"2020-09-13T04:31:41.067Z"},{"product_id":22,"price":"60.00","discount":"7","start":"2020-08-17T22:08:54.055Z","expiry":"2020-09-11T22:08:54.055Z"},{"product_id":23,"price":"11.00","discount":"9","start":"2020-08-19T21:17:42.841Z","expiry":"2020-09-13T21:17:42.841Z"},{"product_id":24,"price":"28.00","discount":"8","start":"2020-08-19T13:46:24.315Z","expiry":"2020-09-13T13:46:24.315Z"},{"product_id":25,"price":"24.00","discount":"8","start":"2020-08-21T10:49:21.264Z","expiry":"2020-09-15T10:49:21.264Z"},{"product_id":26,"price":"75.00","discount":"1","start":"2020-08-21T22:12:56.826Z","expiry":"2020-09-15T22:12:56.826Z"},{"product_id":27,"price":"79.00","discount":"5","start":"2020-08-22T01:21:53.580Z","expiry":"2020-09-16T01:21:53.580Z"},{"product_id":28,"price":"73.00","discount":"1","start":"2020-08-19T23:21:43.706Z","expiry":"2020-09-13T23:21:43.706Z"},{"product_id":29,"price":"95.00","discount":"3","start":"2020-08-21T06:16:00.528Z","expiry":"2020-09-15T06:16:00.528Z"},{"product_id":30,"price":"58.00","discount":"9","start":"2020-08-19T19:42:03.426Z","expiry":"2020-09-13T19:42:03.426Z"},{"product_id":31,"price":"30.00","discount":"6","start":"2020-08-19T09:46:03.217Z","expiry":"2020-09-13T09:46:03.217Z"},{"product_id":32,"price":"58.00","discount":"1","start":"2020-08-17T20:15:20.022Z","expiry":"2020-09-11T20:15:20.022Z"},{"product_id":33,"price":"72.00","discount":"9","start":"2020-08-17T23:02:29.168Z","expiry":"2020-09-11T23:02:29.168Z"},{"product_id":34,"price":"55.00","discount":"2","start":"2020-08-19T10:40:41.081Z","expiry":"2020-09-13T10:40:41.081Z"},{"product_id":35,"price":"96.00","discount":"8","start":"2020-08-21T23:37:48.875Z","expiry":"2020-09-15T23:37:48.875Z"},{"product_id":36,"price":"75.00","discount":"7","start":"2020-08-19T17:55:21.233Z","expiry":"2020-09-13T17:55:21.233Z"},{"product_id":37,"price":"42.00","discount":"6","start":"2020-08-22T01:55:08.997Z","expiry":"2020-09-16T01:55:08.997Z"},{"product_id":38,"price":"90.00","discount":"2","start":"2020-08-22T14:48:26.981Z","expiry":"2020-09-16T14:48:26.981Z"},{"product_id":39,"price":"33.00","discount":"6","start":"2020-08-18T09:25:19.877Z","expiry":"2020-09-12T09:25:19.877Z"},{"product_id":40,"price":"33.00","discount":"7","start":"2020-08-18T08:10:03.577Z","expiry":"2020-09-12T08:10:03.577Z"},{"product_id":41,"price":"50.00","discount":"2","start":"2020-08-18T01:13:47.196Z","expiry":"2020-09-12T01:13:47.196Z"},{"product_id":42,"price":"87.00","discount":"4","start":"2020-08-20T20:45:28.865Z","expiry":"2020-09-14T20:45:28.865Z"},{"product_id":43,"price":"83.00","discount":"6","start":"2020-08-21T22:38:20.181Z","expiry":"2020-09-15T22:38:20.181Z"},{"product_id":44,"price":"29.00","discount":"5","start":"2020-08-21T04:43:41.840Z","expiry":"2020-09-15T04:43:41.840Z"},{"product_id":45,"price":"26.00","discount":"7","start":"2020-08-17T19:19:01.579Z","expiry":"2020-09-11T19:19:01.579Z"},{"product_id":46,"price":"56.00","discount":"9","start":"2020-08-17T21:10:47.798Z","expiry":"2020-09-11T21:10:47.798Z"},{"product_id":47,"price":"93.00","discount":"9","start":"2020-08-21T19:36:09.919Z","expiry":"2020-09-15T19:36:09.919Z"},{"product_id":48,"price":"18.00","discount":"9","start":"2020-08-22T11:09:54.568Z","expiry":"2020-09-16T11:09:54.568Z"},{"product_id":49,"price":"61.00","discount":"8","start":"2020-08-20T16:32:20.384Z","expiry":"2020-09-14T16:32:20.384Z"},{"product_id":50,"price":"82.00","discount":"9","start":"2020-08-18T06:38:53.112Z","expiry":"2020-09-12T06:38:53.112Z"},{"product_id":51,"price":"19.00","discount":"3","start":"2020-08-18T04:37:39.700Z","expiry":"2020-09-12T04:37:39.700Z"},{"product_id":52,"price":"67.00","discount":"1","start":"2020-08-20T15:23:21.909Z","expiry":"2020-09-14T15:23:21.909Z"},{"product_id":53,"price":"48.00","discount":"9","start":"2020-08-21T19:34:30.598Z","expiry":"2020-09-15T19:34:30.598Z"},{"product_id":54,"price":"12.00","discount":"8","start":"2020-08-19T05:16:58.677Z","expiry":"2020-09-13T05:16:58.677Z"},{"product_id":55,"price":"19.00","discount":"2","start":"2020-08-22T04:00:38.241Z","expiry":"2020-09-16T04:00:38.241Z"},{"product_id":56,"price":"79.00","discount":"3","start":"2020-08-21T06:08:56.314Z","expiry":"2020-09-15T06:08:56.314Z"},{"product_id":57,"price":"18.00","discount":"1","start":"2020-08-17T17:21:49.885Z","expiry":"2020-09-11T17:21:49.885Z"},{"product_id":58,"price":"35.00","discount":"2","start":"2020-08-19T07:16:04.262Z","expiry":"2020-09-13T07:16:04.262Z"},{"product_id":59,"price":"34.00","discount":"5","start":"2020-08-18T00:52:46.129Z","expiry":"2020-09-12T00:52:46.129Z"},{"product_id":60,"price":"12.00","discount":"4","start":"2020-08-20T06:10:29.914Z","expiry":"2020-09-14T06:10:29.914Z"},{"product_id":61,"price":"20.00","discount":"2","start":"2020-08-18T10:26:49.064Z","expiry":"2020-09-12T10:26:49.064Z"},{"product_id":62,"price":"25.00","discount":"8","start":"2020-08-17T23:46:19.230Z","expiry":"2020-09-11T23:46:19.230Z"},{"product_id":63,"price":"83.00","discount":"5","start":"2020-08-22T05:26:27.015Z","expiry":"2020-09-16T05:26:27.015Z"},{"product_id":64,"price":"49.00","discount":"4","start":"2020-08-18T21:40:53.590Z","expiry":"2020-09-12T21:40:53.590Z"},{"product_id":65,"price":"28.00","discount":"9","start":"2020-08-20T13:48:40.135Z","expiry":"2020-09-14T13:48:40.135Z"},{"product_id":66,"price":"33.00","discount":"4","start":"2020-08-19T09:52:49.576Z","expiry":"2020-09-13T09:52:49.576Z"},{"product_id":67,"price":"80.00","discount":"9","start":"2020-08-19T15:18:55.025Z","expiry":"2020-09-13T15:18:55.025Z"},{"product_id":68,"price":"71.00","discount":"3","start":"2020-08-18T08:04:52.594Z","expiry":"2020-09-12T08:04:52.594Z"},{"product_id":69,"price":"14.00","discount":"8","start":"2020-08-18T18:34:10.303Z","expiry":"2020-09-12T18:34:10.303Z"},{"product_id":70,"price":"56.00","discount":"8","start":"2020-08-20T23:11:06.731Z","expiry":"2020-09-14T23:11:06.731Z"},{"product_id":71,"price":"20.00","discount":"4","start":"2020-08-19T19:31:15.235Z","expiry":"2020-09-13T19:31:15.235Z"},{"product_id":72,"price":"84.00","discount":"6","start":"2020-08-19T08:24:03.659Z","expiry":"2020-09-13T08:24:03.659Z"},{"product_id":73,"price":"65.00","discount":"1","start":"2020-08-21T22:00:08.354Z","expiry":"2020-09-15T22:00:08.354Z"},{"product_id":74,"price":"76.00","discount":"2","start":"2020-08-21T23:12:07.958Z","expiry":"2020-09-15T23:12:07.958Z"},{"product_id":75,"price":"23.00","discount":"2","start":"2020-08-21T21:26:50.763Z","expiry":"2020-09-15T21:26:50.763Z"},{"product_id":76,"price":"91.00","discount":"5","start":"2020-08-17T20:38:54.455Z","expiry":"2020-09-11T20:38:54.455Z"},{"product_id":77,"price":"56.00","discount":"6","start":"2020-08-18T04:13:38.872Z","expiry":"2020-09-12T04:13:38.872Z"},{"product_id":78,"price":"72.00","discount":"4","start":"2020-08-22T02:01:54.091Z","expiry":"2020-09-16T02:01:54.091Z"},{"product_id":79,"price":"50.00","discount":"9","start":"2020-08-19T23:15:00.680Z","expiry":"2020-09-13T23:15:00.680Z"},{"product_id":80,"price":"44.00","discount":"4","start":"2020-08-20T13:26:51.251Z","expiry":"2020-09-14T13:26:51.251Z"},{"product_id":81,"price":"56.00","discount":"5","start":"2020-08-21T22:53:57.480Z","expiry":"2020-09-15T22:53:57.480Z"},{"product_id":82,"price":"91.00","discount":"2","start":"2020-08-19T15:14:00.158Z","expiry":"2020-09-13T15:14:00.158Z"},{"product_id":83,"price":"98.00","discount":"1","start":"2020-08-18T05:06:13.971Z","expiry":"2020-09-12T05:06:13.971Z"},{"product_id":84,"price":"63.00","discount":"8","start":"2020-08-20T06:56:51.926Z","expiry":"2020-09-14T06:56:51.926Z"},{"product_id":85,"price":"60.00","discount":"8","start":"2020-08-21T15:24:39.539Z","expiry":"2020-09-15T15:24:39.539Z"},{"product_id":86,"price":"49.00","discount":"2","start":"2020-08-19T10:28:25.051Z","expiry":"2020-09-13T10:28:25.051Z"},{"product_id":87,"price":"46.00","discount":"4","start":"2020-08-22T03:00:47.042Z","expiry":"2020-09-16T03:00:47.042Z"},{"product_id":88,"price":"56.00","discount":"1","start":"2020-08-21T16:57:06.398Z","expiry":"2020-09-15T16:57:06.398Z"},{"product_id":89,"price":"70.00","discount":"4","start":"2020-08-22T01:34:41.872Z","expiry":"2020-09-16T01:34:41.872Z"},{"product_id":90,"price":"44.00","discount":"9","start":"2020-08-17T22:04:39.133Z","expiry":"2020-09-11T22:04:39.133Z"},{"product_id":91,"price":"88.00","discount":"7","start":"2020-08-18T05:35:10.885Z","expiry":"2020-09-12T05:35:10.885Z"},{"product_id":92,"price":"30.00","discount":"5","start":"2020-08-18T22:54:51.488Z","expiry":"2020-09-12T22:54:51.488Z"},{"product_id":93,"price":"18.00","discount":"9","start":"2020-08-22T01:15:44.434Z","expiry":"2020-09-16T01:15:44.434Z"},{"product_id":94,"price":"29.00","discount":"9","start":"2020-08-18T17:34:13.253Z","expiry":"2020-09-12T17:34:13.253Z"},{"product_id":95,"price":"66.00","discount":"4","start":"2020-08-22T14:06:47.923Z","expiry":"2020-09-16T14:06:47.923Z"},{"product_id":96,"price":"13.00","discount":"3","start":"2020-08-21T00:56:25.258Z","expiry":"2020-09-15T00:56:25.258Z"},{"product_id":97,"price":"19.00","discount":"7","start":"2020-08-22T09:35:25.472Z","expiry":"2020-09-16T09:35:25.472Z"},{"product_id":98,"price":"16.00","discount":"1","start":"2020-08-19T09:33:43.339Z","expiry":"2020-09-13T09:33:43.339Z"},{"product_id":99,"price":"82.00","discount":"5","start":"2020-08-17T15:03:45.851Z","expiry":"2020-09-11T15:03:45.851Z"},{"product_id":100,"price":"49.00","discount":"9","start":"2020-08-18T14:23:37.096Z","expiry":"2020-09-12T14:23:37.096Z"}]
+[{"product_id":1,"price":"34.00","discount":"9","start":"2020-08-31T06:51:07.255Z","expiry":"2020-09-25T06:51:07.255Z"},{"product_id":2,"price":"95.00","discount":"8","start":"2020-08-29T20:01:26.777Z","expiry":"2020-09-23T20:01:26.777Z"},{"product_id":3,"price":"35.00","discount":"9","start":"2020-08-29T16:17:16.601Z","expiry":"2020-09-23T16:17:16.601Z"},{"product_id":4,"price":"43.00","discount":"8","start":"2020-08-28T05:04:17.297Z","expiry":"2020-09-22T05:04:17.297Z"},{"product_id":5,"price":"53.00","discount":"5","start":"2020-08-30T16:21:33.644Z","expiry":"2020-09-24T16:21:33.644Z"},{"product_id":6,"price":"83.00","discount":"1","start":"2020-08-27T17:22:11.593Z","expiry":"2020-09-21T17:22:11.593Z"},{"product_id":7,"price":"97.00","discount":"7","start":"2020-08-31T14:48:45.134Z","expiry":"2020-09-25T14:48:45.134Z"},{"product_id":8,"price":"59.00","discount":"2","start":"2020-08-28T03:58:00.793Z","expiry":"2020-09-22T03:58:00.793Z"},{"product_id":9,"price":"30.00","discount":"6","start":"2020-08-29T23:51:50.411Z","expiry":"2020-09-23T23:51:50.411Z"},{"product_id":10,"price":"69.00","discount":"6","start":"2020-08-29T00:06:49.204Z","expiry":"2020-09-23T00:06:49.204Z"},{"product_id":11,"price":"51.00","discount":"3","start":"2020-08-27T06:57:00.536Z","expiry":"2020-09-21T06:57:00.536Z"},{"product_id":12,"price":"70.00","discount":"5","start":"2020-08-29T15:13:41.191Z","expiry":"2020-09-23T15:13:41.191Z"},{"product_id":13,"price":"25.00","discount":"1","start":"2020-08-31T06:46:23.627Z","expiry":"2020-09-25T06:46:23.627Z"},{"product_id":14,"price":"42.00","discount":"4","start":"2020-08-30T16:20:54.854Z","expiry":"2020-09-24T16:20:54.854Z"},{"product_id":15,"price":"47.00","discount":"5","start":"2020-08-28T10:06:37.346Z","expiry":"2020-09-22T10:06:37.346Z"},{"product_id":16,"price":"93.00","discount":"4","start":"2020-08-29T07:34:32.985Z","expiry":"2020-09-23T07:34:32.985Z"},{"product_id":17,"price":"92.00","discount":"9","start":"2020-08-28T04:02:18.323Z","expiry":"2020-09-22T04:02:18.323Z"},{"product_id":18,"price":"40.00","discount":"5","start":"2020-09-01T02:01:32.885Z","expiry":"2020-09-26T02:01:32.885Z"},{"product_id":19,"price":"58.00","discount":"5","start":"2020-08-27T14:56:05.984Z","expiry":"2020-09-21T14:56:05.984Z"},{"product_id":20,"price":"65.00","discount":"5","start":"2020-08-27T16:16:49.735Z","expiry":"2020-09-21T16:16:49.735Z"},{"product_id":21,"price":"99.00","discount":"6","start":"2020-09-01T01:20:07.156Z","expiry":"2020-09-26T01:20:07.156Z"},{"product_id":22,"price":"71.00","discount":"9","start":"2020-08-31T18:06:12.849Z","expiry":"2020-09-25T18:06:12.849Z"},{"product_id":23,"price":"78.00","discount":"8","start":"2020-08-28T22:15:40.705Z","expiry":"2020-09-22T22:15:40.705Z"},{"product_id":24,"price":"80.00","discount":"9","start":"2020-08-31T16:58:19.638Z","expiry":"2020-09-25T16:58:19.638Z"},{"product_id":25,"price":"69.00","discount":"1","start":"2020-08-27T17:09:53.011Z","expiry":"2020-09-21T17:09:53.011Z"},{"product_id":26,"price":"87.00","discount":"3","start":"2020-08-30T01:17:19.781Z","expiry":"2020-09-24T01:17:19.781Z"},{"product_id":27,"price":"57.00","discount":"6","start":"2020-08-28T05:07:14.574Z","expiry":"2020-09-22T05:07:14.574Z"},{"product_id":28,"price":"16.00","discount":"1","start":"2020-08-29T18:50:20.307Z","expiry":"2020-09-23T18:50:20.307Z"},{"product_id":29,"price":"54.00","discount":"7","start":"2020-08-31T03:46:28.306Z","expiry":"2020-09-25T03:46:28.306Z"},{"product_id":30,"price":"50.00","discount":"4","start":"2020-08-27T12:56:22.192Z","expiry":"2020-09-21T12:56:22.192Z"},{"product_id":31,"price":"13.00","discount":"1","start":"2020-08-29T23:25:14.427Z","expiry":"2020-09-23T23:25:14.427Z"},{"product_id":32,"price":"18.00","discount":"6","start":"2020-08-28T19:59:20.529Z","expiry":"2020-09-22T19:59:20.529Z"},{"product_id":33,"price":"81.00","discount":"4","start":"2020-08-30T15:54:50.845Z","expiry":"2020-09-24T15:54:50.845Z"},{"product_id":34,"price":"58.00","discount":"7","start":"2020-08-28T08:00:22.815Z","expiry":"2020-09-22T08:00:22.815Z"},{"product_id":35,"price":"58.00","discount":"7","start":"2020-08-30T20:07:27.741Z","expiry":"2020-09-24T20:07:27.741Z"},{"product_id":36,"price":"52.00","discount":"6","start":"2020-08-29T03:51:35.114Z","expiry":"2020-09-23T03:51:35.114Z"},{"product_id":37,"price":"24.00","discount":"8","start":"2020-08-27T09:32:43.583Z","expiry":"2020-09-21T09:32:43.583Z"},{"product_id":38,"price":"46.00","discount":"6","start":"2020-08-31T09:10:43.328Z","expiry":"2020-09-25T09:10:43.328Z"},{"product_id":39,"price":"62.00","discount":"3","start":"2020-08-30T05:09:12.344Z","expiry":"2020-09-24T05:09:12.344Z"},{"product_id":40,"price":"81.00","discount":"1","start":"2020-08-29T13:32:31.334Z","expiry":"2020-09-23T13:32:31.334Z"},{"product_id":41,"price":"100.00","discount":"9","start":"2020-08-27T18:51:59.542Z","expiry":"2020-09-21T18:51:59.542Z"},{"product_id":42,"price":"71.00","discount":"8","start":"2020-08-29T10:34:58.297Z","expiry":"2020-09-23T10:34:58.297Z"},{"product_id":43,"price":"52.00","discount":"2","start":"2020-08-31T05:40:06.606Z","expiry":"2020-09-25T05:40:06.606Z"},{"product_id":44,"price":"43.00","discount":"4","start":"2020-08-29T11:49:47.770Z","expiry":"2020-09-23T11:49:47.770Z"},{"product_id":45,"price":"93.00","discount":"6","start":"2020-08-30T06:39:39.516Z","expiry":"2020-09-24T06:39:39.516Z"},{"product_id":46,"price":"38.00","discount":"8","start":"2020-08-28T19:07:29.501Z","expiry":"2020-09-22T19:07:29.501Z"},{"product_id":47,"price":"73.00","discount":"2","start":"2020-08-29T11:46:24.416Z","expiry":"2020-09-23T11:46:24.416Z"},{"product_id":48,"price":"69.00","discount":"8","start":"2020-08-31T07:30:29.800Z","expiry":"2020-09-25T07:30:29.800Z"},{"product_id":49,"price":"87.00","discount":"5","start":"2020-08-28T07:51:10.784Z","expiry":"2020-09-22T07:51:10.784Z"},{"product_id":50,"price":"44.00","discount":"9","start":"2020-08-31T09:01:15.975Z","expiry":"2020-09-25T09:01:15.975Z"},{"product_id":51,"price":"67.00","discount":"8","start":"2020-08-31T00:10:24.161Z","expiry":"2020-09-25T00:10:24.161Z"},{"product_id":52,"price":"75.00","discount":"8","start":"2020-08-28T03:28:53.238Z","expiry":"2020-09-22T03:28:53.238Z"},{"product_id":53,"price":"95.00","discount":"8","start":"2020-08-31T01:40:13.405Z","expiry":"2020-09-25T01:40:13.405Z"},{"product_id":54,"price":"59.00","discount":"1","start":"2020-08-30T08:26:39.776Z","expiry":"2020-09-24T08:26:39.776Z"},{"product_id":55,"price":"65.00","discount":"2","start":"2020-08-31T15:46:12.170Z","expiry":"2020-09-25T15:46:12.170Z"},{"product_id":56,"price":"72.00","discount":"6","start":"2020-08-28T22:11:04.809Z","expiry":"2020-09-22T22:11:04.809Z"},{"product_id":57,"price":"99.00","discount":"1","start":"2020-08-31T21:07:58.098Z","expiry":"2020-09-25T21:07:58.098Z"},{"product_id":58,"price":"23.00","discount":"9","start":"2020-08-31T20:22:49.134Z","expiry":"2020-09-25T20:22:49.134Z"},{"product_id":59,"price":"30.00","discount":"9","start":"2020-08-27T07:04:38.083Z","expiry":"2020-09-21T07:04:38.083Z"},{"product_id":60,"price":"58.00","discount":"8","start":"2020-08-28T04:50:27.573Z","expiry":"2020-09-22T04:50:27.573Z"},{"product_id":61,"price":"61.00","discount":"6","start":"2020-08-30T00:27:46.861Z","expiry":"2020-09-24T00:27:46.861Z"},{"product_id":62,"price":"58.00","discount":"4","start":"2020-08-27T04:55:05.112Z","expiry":"2020-09-21T04:55:05.112Z"},{"product_id":63,"price":"81.00","discount":"2","start":"2020-08-31T12:04:04.303Z","expiry":"2020-09-25T12:04:04.303Z"},{"product_id":64,"price":"17.00","discount":"8","start":"2020-08-27T21:34:17.830Z","expiry":"2020-09-21T21:34:17.830Z"},{"product_id":65,"price":"66.00","discount":"4","start":"2020-08-31T23:19:11.163Z","expiry":"2020-09-25T23:19:11.163Z"},{"product_id":66,"price":"18.00","discount":"4","start":"2020-08-29T10:21:40.090Z","expiry":"2020-09-23T10:21:40.090Z"},{"product_id":67,"price":"88.00","discount":"7","start":"2020-08-28T10:35:58.509Z","expiry":"2020-09-22T10:35:58.509Z"},{"product_id":68,"price":"55.00","discount":"9","start":"2020-08-28T14:44:02.021Z","expiry":"2020-09-22T14:44:02.021Z"},{"product_id":69,"price":"60.00","discount":"4","start":"2020-08-27T18:39:13.896Z","expiry":"2020-09-21T18:39:13.896Z"},{"product_id":70,"price":"29.00","discount":"4","start":"2020-08-30T05:16:24.198Z","expiry":"2020-09-24T05:16:24.198Z"},{"product_id":71,"price":"56.00","discount":"9","start":"2020-08-27T09:59:08.896Z","expiry":"2020-09-21T09:59:08.896Z"},{"product_id":72,"price":"31.00","discount":"1","start":"2020-08-28T21:10:15.669Z","expiry":"2020-09-22T21:10:15.669Z"},{"product_id":73,"price":"26.00","discount":"7","start":"2020-08-31T19:15:57.400Z","expiry":"2020-09-25T19:15:57.400Z"},{"product_id":74,"price":"28.00","discount":"4","start":"2020-08-30T21:23:33.832Z","expiry":"2020-09-24T21:23:33.832Z"},{"product_id":75,"price":"82.00","discount":"8","start":"2020-08-28T14:45:50.653Z","expiry":"2020-09-22T14:45:50.653Z"},{"product_id":76,"price":"69.00","discount":"1","start":"2020-08-28T06:14:33.061Z","expiry":"2020-09-22T06:14:33.061Z"},{"product_id":77,"price":"67.00","discount":"5","start":"2020-08-30T07:48:57.817Z","expiry":"2020-09-24T07:48:57.817Z"},{"product_id":78,"price":"34.00","discount":"1","start":"2020-08-31T11:51:56.647Z","expiry":"2020-09-25T11:51:56.647Z"},{"product_id":79,"price":"39.00","discount":"8","start":"2020-08-27T16:34:00.196Z","expiry":"2020-09-21T16:34:00.196Z"},{"product_id":80,"price":"82.00","discount":"5","start":"2020-08-27T10:01:11.873Z","expiry":"2020-09-21T10:01:11.873Z"},{"product_id":81,"price":"22.00","discount":"1","start":"2020-08-27T18:19:21.904Z","expiry":"2020-09-21T18:19:21.904Z"},{"product_id":82,"price":"99.00","discount":"4","start":"2020-08-29T18:23:08.801Z","expiry":"2020-09-23T18:23:08.801Z"},{"product_id":83,"price":"15.00","discount":"8","start":"2020-08-31T22:01:43.965Z","expiry":"2020-09-25T22:01:43.965Z"},{"product_id":84,"price":"18.00","discount":"4","start":"2020-08-27T16:10:37.005Z","expiry":"2020-09-21T16:10:37.005Z"},{"product_id":85,"price":"65.00","discount":"5","start":"2020-08-31T09:02:14.502Z","expiry":"2020-09-25T09:02:14.502Z"},{"product_id":86,"price":"71.00","discount":"5","start":"2020-08-28T05:39:30.285Z","expiry":"2020-09-22T05:39:30.285Z"},{"product_id":87,"price":"54.00","discount":"9","start":"2020-08-29T16:55:19.386Z","expiry":"2020-09-23T16:55:19.386Z"},{"product_id":88,"price":"47.00","discount":"5","start":"2020-08-28T22:02:11.583Z","expiry":"2020-09-22T22:02:11.583Z"},{"product_id":89,"price":"43.00","discount":"9","start":"2020-08-28T08:23:26.929Z","expiry":"2020-09-22T08:23:26.929Z"},{"product_id":90,"price":"48.00","discount":"3","start":"2020-08-30T07:19:57.663Z","expiry":"2020-09-24T07:19:57.663Z"},{"product_id":91,"price":"58.00","discount":"1","start":"2020-08-31T18:58:00.531Z","expiry":"2020-09-25T18:58:00.531Z"},{"product_id":92,"price":"23.00","discount":"4","start":"2020-08-30T19:10:13.503Z","expiry":"2020-09-24T19:10:13.503Z"},{"product_id":93,"price":"88.00","discount":"8","start":"2020-08-28T18:51:14.771Z","expiry":"2020-09-22T18:51:14.771Z"},{"product_id":94,"price":"30.00","discount":"3","start":"2020-08-31T12:08:43.471Z","expiry":"2020-09-25T12:08:43.471Z"},{"product_id":95,"price":"34.00","discount":"8","start":"2020-08-31T09:43:03.081Z","expiry":"2020-09-25T09:43:03.081Z"},{"product_id":96,"price":"37.00","discount":"3","start":"2020-08-30T08:54:08.325Z","expiry":"2020-09-24T08:54:08.325Z"},{"product_id":97,"price":"52.00","discount":"8","start":"2020-08-29T15:11:54.642Z","expiry":"2020-09-23T15:11:54.642Z"},{"product_id":98,"price":"98.00","discount":"7","start":"2020-08-29T23:10:31.428Z","expiry":"2020-09-23T23:10:31.428Z"},{"product_id":99,"price":"31.00","discount":"2","start":"2020-08-27T13:01:18.056Z","expiry":"2020-09-21T13:01:18.056Z"},{"product_id":100,"price":"13.00","discount":"1","start":"2020-08-31T01:24:23.060Z","expiry":"2020-09-25T01:24:23.060Z"}]
